+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 0.10.0
 
+### Enhancements
+- **Import Command**
+  - `spout import` now changes variables with all caps `display_name` to use title case instead
+    - Display names that use mixed case are unaffected
+
 ## 0.9.1 (October 14, 2014)
 
 ### Bug Fix